requirements.txt
Flask
python-dotenv
together
Werkzeug